requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
numpy
jinja2
python-dotenv